Move tsdb bwc tests to x-pack/logsdb (#139671)
